Refactory class packages and split class in new class
