+++ src/main/java/com/br/kerberus/urd/entity/LogError.java
@@ -1,4 +1,4 @@
-package com.br.kerberus.urd.log;
+package com.br.kerberus.urd.entity;
 
 import org.springframework.http.HttpStatus;
