+++ src/main/java/com/br/kerberus/urd/entity/LogHttpMessages.java
@@ -1,0 +1,14 @@
+package com.br.kerberus.urd.entity;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LogHttpMessages {
+
+    LogType LogType() default LogType.HTTP_MESSAGES;
+
+}
