+++ src/main/java/com/br/kerberus/urd/entity/LogMessage.java
@@ -3,8 +3,6 @@ package com.br.kerberus.urd.entity;
 import java.util.Date;
 import java.util.Objects;
 
-import com.br.kerberus.urd.log.SystemLog;
-
 public final class LogMessage {
 
 	private String serverName;
