+++ src/main/java/com/br/kerberus/urd/entity/LogMethodCall.java
@@ -1,4 +1,4 @@
-package com.br.kerberus.urd.log;
+package com.br.kerberus.urd.entity;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -9,6 +9,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface LogMethodCall {
 
-    LogType LogType() default LogType.METHOD;
+    LogType LogType() default LogType.METHOD_CALL;
 
 }
