+++ src/main/java/com/br/kerberus/urd/entity/LogMetlhodReturn.java
@@ -1,4 +1,4 @@
-package com.br.kerberus.urd.log;
+package com.br.kerberus.urd.entity;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,7 +7,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LogException {
+public @interface LogMetlhodReturn {
 
-    LogType LogType() default LogType.EXCEPTION;
+    LogType LogType() default LogType.METHOD_RETURN;
 }
