+++ src/main/java/com/br/kerberus/urd/entity/LogType.java
@@ -1,0 +1,6 @@
+package com.br.kerberus.urd.entity;
+
+public enum LogType {
+
+    METHOD_CALL, METHOD_RETURN, HTTP_MESSAGES, EXECUTION_TIME, EXCEPTION, GENERAL, SOAP_MESSAGES, DEBUG
+}
