--- src/main/java/com/br/kerberus/urd/log/LogType.java
@@ -1,6 +1,0 @@
-package com.br.kerberus.urd.log;
-
-public enum LogType {
-
-    METHOD, HTTP_REQUEST, EXECUTION_TIME, EXCEPTION, GENERAL
-}
