+++ src/main/java/com/br/kerberus/urd/service/LogService.java
@@ -1,0 +1,10 @@
+package com.br.kerberus.urd.service;
+
+import org.slf4j.Logger;
+
+public interface LogService {
+
+    public Logger getLog();
+
+    public void setLog(Logger log);
+}
